raw_to_json/requirements.txt
# Kafka client
confluent-kafka==2.3.0

# Configuration parsing
PyYAML==6.0.1

# Logging
python-json-logger==2.0.7
